Renamed sensitiveInformation to correct spelling throughout
